AppStorePaymentManager: refactor
