+++ ios/MullvadVPN/AppStorePaymentManager/AppStorePaymentManagerDelegate.swift
@@ -1,0 +1,18 @@
+//
+//  AppStorePaymentManagerDelegate.swift
+//  AppStorePaymentManagerDelegate
+//
+//  Created by pronebird on 03/09/2021.
+//  Copyright © 2021 Mullvad VPN AB. All rights reserved.
+//
+
+import Foundation
+import StoreKit
+
+protocol AppStorePaymentManagerDelegate: AnyObject {
+
+    /// Return the account token associated with the payment.
+    /// Usually called for unfinished transactions coming back after the app was restarted.
+    func appStorePaymentManager(_ manager: AppStorePaymentManager,
+                                didRequestAccountTokenFor payment: SKPayment) -> String?
+}
