+++ ios/MullvadVPN/Logging/LogFormatting.swift
@@ -1,0 +1,18 @@
+//
+//  LogFormatting.swift
+//  LogFormatting
+//
+//  Created by pronebird on 09/09/2021.
+//  Copyright © 2021 Mullvad VPN AB. All rights reserved.
+//
+
+import Foundation
+
+extension Date {
+    func logFormatDate() -> String {
+        let formatter = DateFormatter()
+        formatter.dateFormat = "dd/MM/yyyy @ HH:mm"
+
+        return formatter.string(from: self)
+    }
+}
